requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
requests==2.31.0

# Computer Vision & ML
ultralytics==8.1.0
torch==2.1.2
torchvision==0.16.2
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.3

# Data Management
roboflow==1.1.19
kaggle==1.6.3
wandb==0.16.2

# Utilities
tqdm==4.66.1
pandas==2.2.0
fiftyone
